rag_service/requirements.txt
fastapi
uvicorn[standard]
pymongo
qdrant-client
openai
python-dotenv
python-multipart
pdfplumber
python-docx
beautifulsoup4
requests
